mywebsite/serializers.py
from rest_framework import serializers
from .models import Project, Hobby


class ProjectSerializer(serializers.ModelSerializer):
    class Meta:
        model = Project
        fields = ('id', 'title', 'description', 'link', 'location', 'duration')

class HobbySerializer(serializers.ModelSerializer):
    class Meta:
        model = Hobby
        fields = ('id', 'title', 'description')
