requirements.txt
asgiref==3.2.3
astroid==2.3.3
automium==0.2.6
automium-web==0.1.1
Click==7.0
dj-config-url==0.1.1
django-cors-headers==3.2.0
djangorestframework==3.11.0
Flask==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
legacy==0.1.6
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.4.4
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
dj-database-url==0.4.2
Django==2.2
gunicorn==19.7.1
psycopg2==2.7.3.2